utils/doc_generator.py
"""
Document generation utilities for the procurement document generator.
This module will handle the generation of documents based on templates and form data.
"""

from docx import Document
import os
import platform
from io import BytesIO

class DocumentGenerator:
    """
    Class for generating procurement documents from templates.
    This will be implemented in future tasks.
    """
    
    def __init__(self, template_path):
        """
        Initialize the document generator with a template path.
        
        Args:
            template_path (str): Path to the document template
        """
        self.template_path = template_path
    
    def generate_docx(self, data):
        """
        Generate a .docx document from the template and data.
        This will be implemented in task 4.3.
        
        Args:
            data (dict): Form data to populate the template
            
        Returns:
            BytesIO: In-memory file-like object containing the generated document
        """
        # This will be implemented in task 4.3
        pass
    
    def generate_pdf(self, data):
        """
        Generate a PDF document from the template and data.
        This will be implemented in tasks 4.4 and 4.5.
        
        Args:
            data (dict): Form data to populate the template
            
        Returns:
            BytesIO: In-memory file-like object containing the generated document
        """
        # This will be implemented in tasks 4.4 and 4.5
        pass