requirements.txt
flask==2.3.3
python-docx==0.8.11
weasyprint==59.0
docx2pdf==0.1.8
python-dateutil==2.8.2
pytest==7.4.0
Jinja2==3.1.3
Werkzeug==2.3.7
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2